+++ spec/filewatch/spec_helper.rb
@@ -1,3 +1,4 @@
+# encoding: utf-8
 require "rspec_sequencing"
 require 'rspec/wait'
 require "logstash/devutils/rspec/spec_helper"
@@ -118,3 +119,8 @@ module FileWatch
       @listeners.clear; end
   end
 end
+
+ENV["LOG_AT"].tap do |level|
+  LogStash::Logging::Logger::configure_logging(level) unless level.nil?
+end
+